requirements.txt
Flask==2.3.3
requests
python-dotenv==1.0.0
langchain
langchain-groq 